fix(chat): show checkout modal when asking Claude from PR description
The Ask Claude button on the PR description did not trigger a branch
checkout modal because usePRChat lacked prState, currentBranch,
prHeadBranch, and origin. Mirrors the handleActionWithCheckout pattern
already used in useActivityChat so that the same checkout flow applies
when the current branch differs from the PR head branch.
