--- frontend/src/domains/PRDetail/hooks/usePRChat.ts
@@ -1,76 +1,0 @@
-import type { PRMeta, ClaudeChatContext } from '@lgtmai/backend/types';
-import { useChatPanel } from '../contexts';
-import { useChatPanelParams } from './useChatPanelParams';
-import { useSessionResume } from './useSessionResume';
-import type { UseClaudeWebSocketReturn } from './useClaudeWebSocket';
-import { ACTION_LABELS } from '../utils/reviewPrompts';
-
-interface UsePRChatOptions {
-  projectId: string;
-  prNumber: number;
-  prMeta: PRMeta;
-  prAuthor: string;
-  prBody: string;
-  workingDir: string;
-  ws: UseClaudeWebSocketReturn;
-}
-
-export function usePRChat({
-  projectId,
-  prNumber,
-  prMeta,
-  prAuthor,
-  prBody,
-  workingDir,
-  ws,
-}: UsePRChatOptions) {
-  const { setTargetContext, setPRContext, setOnExecuteAction } = useChatPanel();
-  const { openActionSelector, openChat } = useChatPanelParams();
-  const { registerResumeSession } = useSessionResume({
-    projectId,
-    prNumber,
-    workingDir,
-    ws,
-  });
-
-  // --- Public API ---
-  const openPRChat = () => {
-    if (ws.status !== 'connected') ws.connect();
-    ws.clearMessages();
-
-    setTargetContext({ type: 'pr', author: prAuthor, body: prBody, prNumber });
-    setPRContext({ projectId, prNumber });
-    registerResumeSession();
-
-    setOnExecuteAction((actionId: string, customPrompt?: string) => {
-      const label = ACTION_LABELS[actionId] || customPrompt || actionId;
-      ws.addUserMessage(label);
-
-      const chatContext: ClaudeChatContext = {
-        projectId,
-        targetType: 'PR',
-        targetNumber: prNumber,
-        scopeType: 'PR',
-        scopeTargetId: '',
-        title: label,
-      };
-
-      openChat();
-      ws.execute(
-        {
-          type: 'command',
-          command: actionId as 'review' | 'explain' | 'custom',
-          context: { type: 'pr', prMeta },
-          ...(customPrompt ? { customPrompt } : {}),
-        },
-        workingDir,
-        { executionMode: 'bypassPermissions' },
-        chatContext
-      );
-    });
-
-    openActionSelector('pr', `pr-${prNumber}`);
-  };
-
-  return { openPRChat };
-}
